Issue #2031, moved more methods, refactoring
